Bug fixes Enemies lose their shield at the end of player turn and not at the start of their turns (allows for enemies to shield each other). When an enemy dies it now sets its collision to None (prevents enemies to block ray tracing when they're dead).
When viewing cards from a pile that just had cards added to it, cards
used to become invisible. This is dued to the Pile Logic which hides
cards that are bellow the first visible.
-> Now the Card Collection view sets its cards to visible on the tick.
Since a card collection view only exists when it is called it should be
fine in term of performance.